+++ test/index.js
@@ -5,5 +5,6 @@ load('./test-frame.html');
 load('./test-root.html');
 load('./test-tap.html');
 load('./test-common.html');
+load('./test-assert.html');
 load('./test-boot.html');
 load('./test-scratch.html');